Factored code out of Translator class into new PathSolver class.
Original class was holding all methods for translating p4pktgen graph
paths into z3 constraints, while also managing a z3 solver instance
tracking the path under consideration.  These two tasks are conceptually
different and splitting the class makes work easier.
